+++ project-5/components/layout/mobile-header.tsx
@@ -6,7 +6,7 @@ import { useRouter } from 'next/router'
 import { motion } from 'framer-motion'
 
 import { MOBILE_HEADER_VARIANTS, ROUTES } from 'common/constants'
-import LangToggler from '~components/lang-toggler'
+import LangToggler from '~components/language'
 import MenuItem from '~components/navigation/menu-item'
 import MenuToggler from '~components/navigation/menu-toggler'
 import ThemeToggler from '~components/theme-toggler'
